==> src/madrid/01-madrid-airbnb-eda.ipynb <==
__Máster en Data Science UAH__
# Tasador de viviendas de alquiler vacacional en Madrid
# Notebook #1 - Análisis exploratorio

**Alumno:** Héctor Mateos Oblanca  
**Tutor:** Daniel Rodríguez Pérez


<h1>Contenidos<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Limpieza-y-preparación-de-datos" data-toc-modified-id="Limpieza-y-preparación-de-datos-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Limpieza y preparación de datos</a></span><ul class="toc-item"><li><span><a href="#Carga-de-datos" data-toc-modified-id="Carga-de-datos-1.1"><span class="toc-item-num">1.1&nbsp;&nbsp;</span>Carga de datos</a></span><ul class="toc-item"><li><span><a href="#Dataset-principal-Airbnb" data-toc-modified-id="Dataset-principal-Airbnb-1.1.1"><span class="toc-item-num">1.1.1&nbsp;&nbsp;</span>Dataset principal Airbnb</a></span></li><li><span><a href="#Dataset-de-evolución-precios-de-alquiler-de-Idealista" data-toc-modified-id="Dataset-de-evolución-precios-de-alquiler-de-Idealista-1.1.2"><span class="toc-item-num">1.1.2&nbsp;&nbsp;</span>Dataset de evolución precios de alquiler de Idealista</a></span></li><li><span><a href="#Dataset-de-paradas-de-Metro-y-Cercanías" data-toc-modified-id="Dataset-de-paradas-de-Metro-y-Cercanías-1.1.3"><span class="toc-item-num">1.1.3&nbsp;&nbsp;</span>Dataset de paradas de Metro y Cercanías</a></span></li><li><span><a href="#Carga-de-mapas" data-toc-modified-id="Carga-de-mapas-1.1.4"><span class="toc-item-num">1.1.4&nbsp;&nbsp;</span>Carga de mapas</a></span></li></ul></li><li><span><a href="#Descarte-inicial-de-características" data-toc-modified-id="Descarte-inicial-de-características-1.2"><span class="toc-item-num">1.2&nbsp;&nbsp;</span>Descarte inicial de características</a></span></li><li><span><a href="#Armonización-de-barrios-y-distritos" data-toc-modified-id="Armonización-de-barrios-y-distritos-1.3"><span class="toc-item-num">1.3&nbsp;&nbsp;</span>Armonización de barrios y distritos</a></span></li><li><span><a href="#Características-de-tipo-fecha" data-toc-modified-id="Características-de-tipo-fecha-1.4"><span class="toc-item-num">1.4&nbsp;&nbsp;</span>Características de tipo fecha</a></span></li><li><span><a href="#Características-dinerarias" data-toc-modified-id="Características-dinerarias-1.5"><span class="toc-item-num">1.5&nbsp;&nbsp;</span>Características dinerarias</a></span></li><li><span><a href="#Características-binarias" data-toc-modified-id="Características-binarias-1.6"><span class="toc-item-num">1.6&nbsp;&nbsp;</span>Características binarias</a></span></li><li><span><a href="#Adaptación-de-equipamientos" data-toc-modified-id="Adaptación-de-equipamientos-1.7"><span class="toc-item-num">1.7&nbsp;&nbsp;</span>Adaptación de equipamientos</a></span></li><li><span><a href="#Adaptación-de-las-verificaciones-del-anfitrión" data-toc-modified-id="Adaptación-de-las-verificaciones-del-anfitrión-1.8"><span class="toc-item-num">1.8&nbsp;&nbsp;</span>Adaptación de las verificaciones del anfitrión</a></span></li><li><span><a href="#Adaptación-de-la-licencia" data-toc-modified-id="Adaptación-de-la-licencia-1.9"><span class="toc-item-num">1.9&nbsp;&nbsp;</span>Adaptación de la licencia</a></span></li><li><span><a href="#Nuevas-características-calculadas" data-toc-modified-id="Nuevas-características-calculadas-1.10"><span class="toc-item-num">1.10&nbsp;&nbsp;</span>Nuevas características calculadas</a></span><ul class="toc-item"><li><span><a href="#Tiempo-de-actividad" data-toc-modified-id="Tiempo-de-actividad-1.10.1"><span class="toc-item-num">1.10.1&nbsp;&nbsp;</span>Tiempo de actividad</a></span></li><li><span><a href="#Ingresos-por-estancia-y-nivel-de-ocupación" data-toc-modified-id="Ingresos-por-estancia-y-nivel-de-ocupación-1.10.2"><span class="toc-item-num">1.10.2&nbsp;&nbsp;</span>Ingresos por estancia y nivel de ocupación</a></span></li></ul></li><li><span><a href="#Valores-extremos-(outliers)" data-toc-modified-id="Valores-extremos-(outliers)-1.11"><span class="toc-item-num">1.11&nbsp;&nbsp;</span>Valores extremos (<em>outliers</em>)</a></span><ul class="toc-item"><li><span><a href="#Tipo-de-propiedad" data-toc-modified-id="Tipo-de-propiedad-1.11.1"><span class="toc-item-num">1.11.1&nbsp;&nbsp;</span>Tipo de propiedad</a></span></li><li><span><a href="#Viviendas-sin-reviews" data-toc-modified-id="Viviendas-sin-reviews-1.11.2"><span class="toc-item-num">1.11.2&nbsp;&nbsp;</span>Viviendas sin reviews</a></span></li><li><span><a href="#Precios-por-huésped-extremos" data-toc-modified-id="Precios-por-huésped-extremos-1.11.3"><span class="toc-item-num">1.11.3&nbsp;&nbsp;</span>Precios por huésped extremos</a></span></li><li><span><a href="#Número-mínimo-de-noches-extremo" data-toc-modified-id="Número-mínimo-de-noches-extremo-1.11.4"><span class="toc-item-num">1.11.4&nbsp;&nbsp;</span>Número mínimo de noches extremo</a></span></li><li><span><a href="#Número-de-habitaciones-extremo" data-toc-modified-id="Número-de-habitaciones-extremo-1.11.5"><span class="toc-item-num">1.11.5&nbsp;&nbsp;</span>Número de habitaciones extremo</a></span></li></ul></li></ul></li><li><span><a href="#Análisis-exploratorio" data-toc-modified-id="Análisis-exploratorio-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Análisis exploratorio</a></span><ul class="toc-item"><li><span><a href="#Distribución-del-precio" data-toc-modified-id="Distribución-del-precio-2.1"><span class="toc-item-num">2.1&nbsp;&nbsp;</span>Distribución del precio</a></span></li><li><span><a href="#Resumen-del-resto-de-características" data-toc-modified-id="Resumen-del-resto-de-características-2.2"><span class="toc-item-num">2.2&nbsp;&nbsp;</span>Resumen del resto de características</a></span></li><li><span><a href="#Correlaciones" data-toc-modified-id="Correlaciones-2.3"><span class="toc-item-num">2.3&nbsp;&nbsp;</span>Correlaciones</a></span></li><li><span><a href="#Licencia" data-toc-modified-id="Licencia-2.4"><span class="toc-item-num">2.4&nbsp;&nbsp;</span>Licencia</a></span></li><li><span><a href="#Tipos-de-propiedad-y-relación-con-precio-por-huésped" data-toc-modified-id="Tipos-de-propiedad-y-relación-con-precio-por-huésped-2.5"><span class="toc-item-num">2.5&nbsp;&nbsp;</span>Tipos de propiedad y relación con precio por huésped</a></span></li><li><span><a href="#Tipos-de-habitaciones-y-relación-con-precio-por-huésped" data-toc-modified-id="Tipos-de-habitaciones-y-relación-con-precio-por-huésped-2.6"><span class="toc-item-num">2.6&nbsp;&nbsp;</span>Tipos de habitaciones y relación con precio por huésped</a></span></li><li><span><a href="#Relación-entre-número-y-puntuación-de-las-reviews" data-toc-modified-id="Relación-entre-número-y-puntuación-de-las-reviews-2.7"><span class="toc-item-num">2.7&nbsp;&nbsp;</span>Relación entre número y puntuación de las reviews</a></span></li><li><span><a href="#Relación-entre-política-de-cancelación-y-número-de-reviews-mensuales" data-toc-modified-id="Relación-entre-política-de-cancelación-y-número-de-reviews-mensuales-2.8"><span class="toc-item-num">2.8&nbsp;&nbsp;</span>Relación entre política de cancelación y número de reviews mensuales</a></span></li><li><span><a href="#Relación-entre-número-mínimo-de-noches-y-número-de-reviews-mensuales" data-toc-modified-id="Relación-entre-número-mínimo-de-noches-y-número-de-reviews-mensuales-2.9"><span class="toc-item-num">2.9&nbsp;&nbsp;</span>Relación entre número mínimo de noches y número de reviews mensuales</a></span></li><li><span><a href="#Relación-entre-tiempo-de-respuesta-y-número-de-reviews-mensuales" data-toc-modified-id="Relación-entre-tiempo-de-respuesta-y-número-de-reviews-mensuales-2.10"><span class="toc-item-num">2.10&nbsp;&nbsp;</span>Relación entre tiempo de respuesta y número de reviews mensuales</a></span></li><li><span><a href="#Tasa-de-limpieza" data-toc-modified-id="Tasa-de-limpieza-2.11"><span class="toc-item-num">2.11&nbsp;&nbsp;</span>Tasa de limpieza</a></span></li><li><span><a href="#Número-de-alojamientos-por-barrio" data-toc-modified-id="Número-de-alojamientos-por-barrio-2.12"><span class="toc-item-num">2.12&nbsp;&nbsp;</span>Número de alojamientos por barrio</a></span></li><li><span><a href="#Relación-entre-precio-por-húesped-y-localización" data-toc-modified-id="Relación-entre-precio-por-húesped-y-localización-2.13"><span class="toc-item-num">2.13&nbsp;&nbsp;</span>Relación entre precio por húesped y localización</a></span></li><li><span><a href="#Relación-entre-precio-por-huésped-y-notas-de-review" data-toc-modified-id="Relación-entre-precio-por-huésped-y-notas-de-review-2.14"><span class="toc-item-num">2.14&nbsp;&nbsp;</span>Relación entre precio por huésped y notas de review</a></span></li><li><span><a href="#Relación-entre-precio-por-huésped-y-número-de-reviews-mensuales" data-toc-modified-id="Relación-entre-precio-por-huésped-y-número-de-reviews-mensuales-2.15"><span class="toc-item-num">2.15&nbsp;&nbsp;</span>Relación entre precio por huésped y número de reviews mensuales</a></span></li><li><span><a href="#Review-de-localización" data-toc-modified-id="Review-de-localización-2.16"><span class="toc-item-num">2.16&nbsp;&nbsp;</span>Review de localización</a></span></li><li><span><a href="#Distribución-del-precio-por-huésped" data-toc-modified-id="Distribución-del-precio-por-huésped-2.17"><span class="toc-item-num">2.17&nbsp;&nbsp;</span>Distribución del precio por huésped</a></span></li><li><span><a href="#Relación-entre-habitaciones-y-baños" data-toc-modified-id="Relación-entre-habitaciones-y-baños-2.18"><span class="toc-item-num">2.18&nbsp;&nbsp;</span>Relación entre habitaciones y baños</a></span></li><li><span><a href="#Relación-entre-ingresos-por-estancia-y-capacidad" data-toc-modified-id="Relación-entre-ingresos-por-estancia-y-capacidad-2.19"><span class="toc-item-num">2.19&nbsp;&nbsp;</span>Relación entre ingresos por estancia y capacidad</a></span></li><li><span><a href="#Distribución-de-ingresos-por-estancia-según-el-distrito" data-toc-modified-id="Distribución-de-ingresos-por-estancia-según-el-distrito-2.20"><span class="toc-item-num">2.20&nbsp;&nbsp;</span>Distribución de ingresos por estancia según el distrito</a></span></li><li><span><a href="#Relación-entre-barrio-y-precio-por-huésped" data-toc-modified-id="Relación-entre-barrio-y-precio-por-huésped-2.21"><span class="toc-item-num">2.21&nbsp;&nbsp;</span>Relación entre barrio y precio por huésped</a></span></li><li><span><a href="#Relación-entre-distrito-y-precio-medio-de-alquiler-por-m2" data-toc-modified-id="Relación-entre-distrito-y-precio-medio-de-alquiler-por-m2-2.22"><span class="toc-item-num">2.22&nbsp;&nbsp;</span>Relación entre distrito y precio medio de alquiler por m2</a></span></li><li><span><a href="#Relación-entre-precio-por-huésped-y-precio-medio-de-alquiler-por-m2" data-toc-modified-id="Relación-entre-precio-por-huésped-y-precio-medio-de-alquiler-por-m2-2.23"><span class="toc-item-num">2.23&nbsp;&nbsp;</span>Relación entre precio por huésped y precio medio de alquiler por m2</a></span></li><li><span><a href="#Relación-entre-precio-por-huésped-e-ingresos-por-estancia-media" data-toc-modified-id="Relación-entre-precio-por-huésped-e-ingresos-por-estancia-media-2.24"><span class="toc-item-num">2.24&nbsp;&nbsp;</span>Relación entre precio por huésped e ingresos por estancia media</a></span></li><li><span><a href="#Equipamiento-de-la-vivienda" data-toc-modified-id="Equipamiento-de-la-vivienda-2.25"><span class="toc-item-num">2.25&nbsp;&nbsp;</span>Equipamiento de la vivienda</a></span></li></ul></li><li><span><a href="#Exportación-dataset" data-toc-modified-id="Exportación-dataset-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Exportación dataset</a></span></li></ul></div>

## Limpieza y preparación de datos

import math
import json
import pandas as pd
import pandas_profiling
import numpy as np
import matplotlib.pyplot as plt
import plotly.express as px
import chart_studio.plotly as py
import plotly.graph_objs as go
import plotly.figure_factory as ff
from plotly.colors import n_colors
from plotly.offline import iplot, init_notebook_mode
from plotly.subplots import make_subplots

init_notebook_mode(connected=True)

%run ../utils.py

### Carga de datos

#### Dataset principal Airbnb

city = 'madrid'
month = '201909'
filename_in = '../data/' + city + '-' + month + '-listings.csv'
filename_out = '../data/' + city + '-' + month + '-listings-CLEAN.csv'

df = pd.read_csv(filename_in, low_memory=False)
df.info()

#### Dataset de evolución precios de alquiler de Idealista

dfi = pd.read_csv('../geo/' + city + '.idealista-precio-alquiler-distritos.csv', sep=';')
dfi.info()

#### Dataset de paradas de Metro y Cercanías

dfpt = pd.read_csv('../geo/' + city + '.metro-cercanias-stops.csv')
dfpt.info()

#### Carga de mapas

with open('../geo/' + city + '.districts.geojson') as f:
    city_districts = fix_geojson(json.load(f))
    
with open('../geo/' + city + '.neighbourhoods.geojson') as f:
    city_nb = fix_geojson(json.load(f))

### Descarte inicial de características

En el descarte inicial se desechan las características que se consideran menos útiles para el análisis exploratorio. Tras el análisis exploratorio habrá otro descarte de características que si bien habían sido interesantes para el exploratorio no lo serán para el modelado de la solución.

__Información descriptiva en formato de texto libre__: summary, space, description, name, street, experiences_offered, neighborhood_overview, notes, transit, access, Interaction, house_rules, host_name, host_location, host_about


__Campos que, por lo general, no están informados__: square_feet, weekly_price, monthly_price, beds, jurisdiction_names, host_acceptance_rate.


__Campos que contienen URLs__: thumbnail_url, medium_url, picture_url, xl_picture_url, listing_url, host_url, host_thumbnail_url, host_picture_url.


__Campos con cardinalidad 1 (todos iguales)__: city, state, market, smart_location, country_code, country, last_scraped, requires_license, is_business_travel_ready.


__Campos con cardinalidad máxima (identificadores únicos)__: id, scrape_id, host_id


__Campos dependientes del contexto temporal (datos extraídos del calendario de ocupación tomado en un momento determinado)__:  calendar_updated, has_availability, availability_30, availability_60, availability_90, availability_365, calendar_last_scraped.


__Campos directamente dependientes del anfitrión (dado que el estudio pretende ser de utilidad para viviendas en general que pueden o no estar dadas de alta en Airbnb, se obvian las características relativas a los anfitriones de Airbnb)__: host_since, host_response_rate, host_is_superhost, host_neighbourhood, host_listings_count, host_total_listings_count, host_has_profile_pic, host_identity_verified, calculated_host_listings_count, calculated_host_listings_count_entire_homest, calculated_host_listings_count_private_rooms, calculated_host_listings_count_shared_rooms, require_guest_profile_picture, require_guest_phone_verification, host_verifications.


__Otras circunstancias__:
* neighbourhood: campo no unificado. Se utilizará neighbourhood_cleansed.
* zipcode: división administrativa con poco valor.
* is_location_exact: se dispone de latitud y longitud.
* bed_type: en casi el 99% de las viviendas hay cama real o hay varias camas con lo que no es demasiado descriptivo.
* minimum_nights, maximum_nights, minimum_minimum_nights, maximum_minimum_nights, minimum_maximum_nights, maximum_maximum_nights: se utilizará minimum_nights_avg_ntm y maximum_nights_avg_ntm por simplicidad.


useless_cols = [
    'id',
    'listing_url',
    'scrape_id',
    'last_scraped',
    'name',
    'summary',
    'space',
    'description',
    'experiences_offered',
    'neighborhood_overview',
    'notes',
    'transit',
    'access',
    'interaction',
    'house_rules',
    'thumbnail_url',
    'medium_url',
    'picture_url',
    'xl_picture_url',
    'host_id',
    'host_url',
    'host_name',
    'host_since',
    'host_location',
    'host_about',
    'host_response_rate',
    'host_acceptance_rate',
    'host_is_superhost',
    'host_thumbnail_url',
    'host_picture_url',
    'host_neighbourhood',
    'host_listings_count',
    'host_total_listings_count',
    'host_has_profile_pic',
    'host_identity_verified',
    'host_verifications',
    'street',
    'neighbourhood',
    'city',
    'state',
    'zipcode',
    'market',
    'smart_location',
    'country_code',
    'country',
    'is_location_exact',
    'square_feet',
    'weekly_price',
    'monthly_price',
    'beds',
    'bed_type',
    'minimum_nights',
    'maximum_nights',
    'minimum_minimum_nights',
    'maximum_minimum_nights',
    'minimum_maximum_nights',
    'maximum_maximum_nights',
    'calendar_updated',
    'has_availability',
    'availability_30',
    'availability_60',
    'availability_90',
    'availability_365',
    'calendar_last_scraped',
    'requires_license',
    'jurisdiction_names',
    'is_business_travel_ready',
    'require_guest_profile_picture',
    'require_guest_phone_verification',
    'calculated_host_listings_count',
    'calculated_host_listings_count_entire_homes',
    'calculated_host_listings_count_private_rooms',
    'calculated_host_listings_count_shared_rooms'
]

df.drop(useless_cols, axis=1, inplace=True)

### Armonización de barrios y distritos

Se unifican los nombres de los distritos de los distintos datasets para permitir join entre varios datasets cuando sea necesario.
Para simplificar, se eliminan caracteres especiales como los acentos y se cambia el nombre de la columnas de barrio y distrito.

short_names = np.sort(dfi['Localización'].unique())

def unify_district(n):
    for sn in short_names:
        if sn in n:
            return sn
    
df['district'] = df['neighbourhood_group_cleansed'].apply(lambda x: unify_district(x))
df['district'] = df['district'].apply(lambda x: remove_accents(x))
dfi['Localización'] = dfi['Localización'].apply(lambda x: remove_accents(x))
df.drop(['neighbourhood_group_cleansed'], axis=1, inplace=True)

df['neighbourhood'] = df['neighbourhood_cleansed']
df['neighbourhood'] = df['neighbourhood'].apply(lambda x: remove_accents(x))
df.drop(['neighbourhood_cleansed'], axis=1, inplace=True)

### Características de tipo fecha

Se realiza la conversión de formato texto a formato fecha para facilitar posteriores operaciones.

df['first_review'] = pd.to_datetime(df['first_review'], format='%Y-%m-%d')
df['last_review'] = pd.to_datetime(df['last_review'], format='%Y-%m-%d')

### Características dinerarias

Se realiza la conversión de formato texto en dólares a formato numérico en euros.

dollar_to_euro_rate = 0.9

dollar_cols = [
    'price',
    'security_deposit', 
    'cleaning_fee', 
    'extra_people'
]

for col in dollar_cols:
    if col in df.columns:
        df[col].fillna('$0', inplace=True)
        df[col] = df[col].apply(lambda x: clean_price_dollar(x))
        df[col] = df[col].astype(float)
        df[col] = df[col] * dollar_to_euro_rate
        df[col] = df[col].round(2)

# Dataset Idealista
df_price_rent_m2 = dfi[['Localización', 'Precio m2 oct 2019']].copy()
df_price_rent_m2.columns = ['district', 'price_rent_eur_m2']
df_price_rent_m2['price_rent_eur_m2'] = df_price_rent_m2['price_rent_eur_m2'].apply(lambda x: clean_price_eur(x))
df_price_rent_m2['price_rent_eur_m2'] = df_price_rent_m2['price_rent_eur_m2'].astype(float)

### Características binarias

Se realiza la conversión de formato texto (t para true, f para false) a formato numérico (0 para false, 1 para true).

bin_cols = [
    'host_is_superhost',
    'host_identity_verified', 
    'is_location_exact',
    'has_availability',
    'instant_bookable'
]
    
for col in bin_cols:
    if col in df.columns:
        df[col] = df[col].apply(lambda x: get_bin_value_by_char(x))
        df[col] = df[col].astype(int)

### Adaptación de equipamientos

Se realiza la conversión de un formato no estructurado extrayendo del texto los tipos de equipamientos más relevantes que declara cada vivienda y creando una característica binaria para cada tipo de equipamiento. Esto además simplificará la gestión de variables categóricas en el modelado y facilitará en algunos aspectos la exploración.

df['amenities'] = df['amenities'].str.lower()
amenities_dict = {}

def collect_amenities(str):
    str = str.replace('{', '').replace('}', '').replace('\"', '').strip()
    word_list = str.split(",")
    for w in word_list:
        w = w.strip()
        amenities_dict[w] = amenities_dict.get(w, 0) + 1
    
df['amenities'].apply(lambda x: collect_amenities(x))
amenities_dict = sorted(amenities_dict.items(), key=lambda x: x[1], reverse=True)

top_amenities = [
    'wifi', 
    'essentials', 
    'kitchen', 
    'heating', 
    'washer', 
    'hangers', 
    'tv', 
    'hair dryer', 
    'iron', 
    'shampoo',
    'laptop friendly workspace',
    'air conditioning', 
    'hot water',
    'elevator',
    'refrigerator',
    'dishes and silverware',
    'microwave',
    'bed linens',
    'no stairs or steps to enter',
    'coffee maker',
    'cooking basics',
    'family/kid friendly',
    'long term stays allowed',
    'first aid kit',
    'oven',
    'stove'
]

for v in top_amenities:
    new_col_name = 'has_' + v.replace(' ', '_')
    df[new_col_name] = df['amenities'].apply(lambda x: contains_bin_value(v, x))
    df[new_col_name] = df[new_col_name].astype(int)
    
df.drop(['amenities'], axis=1, inplace=True)

### Adaptación de las verificaciones del anfitrión

Se realiza la conversión de un formato no estructurado extrayendo del texto los tipos de verificaciones más relevantes que declara cada anfitrión y creando una característica binaria para cada tipo de verificación. Esto además simplificará la gestión de variables categóricas en el modelado y facilitará en algunos aspectos la exploración.

"""
df['host_verifications'] = df['host_verifications'].str.lower()
hverif_dict = {}

def collect_host_verifications(str):
    str = str.replace('[', '').replace(']', '').replace('\"', '').replace('\'', '').strip()
    word_list = str.split(",")
    for w in word_list: 
        w = w.strip()
        hverif_dict[w] = hverif_dict.get(w, 0) + 1
    
df['host_verifications'].apply(lambda x: collect_host_verifications(x))
hverif_dict = sorted(hverif_dict.items(), key=lambda x: x[1], reverse=True)

top_verification_modes = [
    'phone',
    'email',
    'government_id',
    'reviews',
    'jumio',
    'offline_government_id',
    'selfie',
    'identity_manual',
    'facebook',
    'work_email',
    'google'
]

for v in top_verification_modes:
    new_col_name = 'host_verified_by_' + v.replace(' ', '_')
    df[new_col_name] = df['host_verifications'].apply(lambda x: contains_bin_value(v, x))
    df[new_col_name] = df[new_col_name].astype(int)
    
df.drop(['host_verifications'], axis=1, inplace=True)
"""

### Adaptación de la licencia
Lo relevante para el estudio es si tiene licencia o no, sin importar el identificador.

df['license'].fillna(0, inplace=True)
df['has_license'] = df['license'].apply(lambda x: 1 if x != 0 else 0)
df.drop(['license'], axis=1, inplace=True)

### Nuevas características calculadas

Se van a generar una serie de características nuevas para enriquecer el estudio descriptivo.

#### Tiempo de actividad

Se trata del período que la vivienda lleva explotándose de forma efectiva y se estima como el tiempo transcurrido en meses desde la primera estancia hasta la última.

df['activity_months'] = (df['last_review'] - df['first_review']) / np.timedelta64(1, 'M')
df['activity_months'].fillna(0, inplace=True)

#### Ingresos por estancia y nivel de ocupación

El precio por noche (_price_) no es un indicador definitivo de los ingresos por estancia porque el precio de una estancia viene determinada por otros factores adicionales:
* Nº de huéspedes (algunos están incluidos en la tarifa price (_guests-included_) y el resto pagan una tarifa extra por noche (_extra-people_)
* Tasas de limpieza (_cleaning-fee_)
* Nº de noches
* Nº de huéspedes (_accomodates_) como capacidad máxima porque beds no está informado en muchos de los casos

Por ello también se puede calcular una variable que estime el coste total de una estancia media tomando como media de número de noches 3.7 y como número de huéspedes el valor medio entre el mínimo y el máximo de ocupantes de la vivienda.

Fuente: [Informe de actividad económica de Airbnb en la ciudad de Madrid](https://press.airbnb.com/wp-content/uploads/sites/4/2019/03/2019-Madrid-Economic-Activity-Report.pdf)

df['cleaning_fee'].fillna(0, inplace=True)
df['extra_people'].fillna(0, inplace=True)
df['guests_included'].fillna(0, inplace=True)
df['accommodates'].fillna(0, inplace=True)

avg_days = 3.7

df['income_med_occupation'] = df.apply(
    lambda r: calculate_income_med_occupation(
        r['price'], 
        r['cleaning_fee'], 
        r['accommodates'], 
        r['extra_people'], 
        r['guests_included'], 
        avg_days), 
    axis=1
)

df['price_med_occupation_per_accommodate'] = df.apply(
    lambda r: calculate_price_med_occupation_per_accommodate(
        r['price'], 
        r['cleaning_fee'], 
        r['accommodates'], 
        r['extra_people'], 
        r['guests_included'], 
        avg_days), 
    axis=1
)

### Valores extremos (_outliers_)

#### Tipo de propiedad
Se excluyen propiedades como los hoteles que ya de por sí son negocios hosteleros ya que el estudio pretende ceñirse a viviendas. 

outliers_idx = df[~df['property_type'].isin(['Apartment', 'House', 'Chalet', 'Condominium', 'Loft'])].index
remove_outliers(df, outliers_idx, debug_col='property_type')

outliers_idx = df[df['room_type'].isin(['Hotel room'])].index
remove_outliers(df, outliers_idx, debug_col='room_type')

#### Viviendas sin reviews
Las viviendas sin reviews, ya que no hay ningún indicio para saber si llevan o no tiempo publicadas, se excluyen del estudio.

df['number_of_reviews'].fillna(0, inplace=True)
outliers_idx = df[df['number_of_reviews'] < 2]['number_of_reviews'].index.tolist()
remove_outliers(df, outliers_idx, debug_col='number_of_reviews')

#### Precios por huésped extremos

Se aplica arbitrariamente una regla de rango intercuartil (IQR) para excluir precios extremos.

outliers_idx = get_outliers_iqr(df['price_med_occupation_per_accommodate'], 4.5)[0]
remove_outliers(df, outliers_idx, debug_col='price_med_occupation_per_accommodate')

#### Número mínimo de noches extremo
Hay viviendas anunciadas cuyo mínimo de noches por estancia es tan elevado (meses, un año o varios años) que se puede deducir que su objetivo o su intención es un alquier de tipo residencial estable, no vacacional.

Se excluyen de forma arbitraria los pisos cuyo número mínimo de noches supere las 70 por considerarlo un alquiler no vacacional.

outliers_idx = df[df['minimum_nights_avg_ntm'] > 70].index
remove_outliers(df, outliers_idx, debug_col='minimum_nights_avg_ntm')

#### Número de habitaciones extremo

Se aplica arbitrariamente una regla de rango intercuartil (IQR) para excluir valores extremos.

df['bedrooms'].fillna(0, inplace=True)
outliers_idx = get_outliers_iqr(df['bedrooms'], 25)[0]
remove_outliers(df, outliers_idx, debug_col='bedrooms')

## Análisis exploratorio

Las características que a priori son las más importantes en el problema que se plantea son los precios y las reviews, por la relación directa en los ingresos del anfitrión y en los costes para el huésped.

Las reviews constituyen la manera más cercana de estimar el nivel de ocupación de la vivienda a falta de datos explícitos al respecto como podrían ser las duraciones de las estancias o el número de huéspedes que definen cada estancia.

En los siguientes apartados se analizan diferentes relaciones entre las múltiples características de los alojamientos, siendo especialmente relevantes los precios y las reviews como se comentaba.

### Distribución del precio

El precio _price_ es la variable objetivo del estudio luego es conveniente ver de partida cómo está distribuido. 

Entre 20 y 100 euros por noche se concentra la gran mayoría de alojamientos.

fig = px.histogram(df, x="price", nbins=40)
fig.show()

fig = ff.create_distplot([df['price']], ['price'], bin_size=[25])
fig.show()

df[['price']].describe()

### Resumen del resto de características

pandas_profiling.ProfileReport(df)

### Correlaciones

En los mapas de correlación se pueden obtener algunos indicios interesantes de relaciones entre características a parte de las relaciones evidentes por familias como por ejemplo la familia de los precios o la familia de las reviews.

def print_corr_map(df, height=None):
    corrs = df.corr()
    fig = go.Figure(
        data=go.Heatmap(
            z=corrs.values,
            x=list(corrs.columns),
            y=list(corrs.index),
            showscale=True
        )
    )
    
    if height:
        fig.update_layout(height=height)
        
    fig.show()

key_cols = [
    'price',
    'price_med_occupation_per_accommodate',
    'income_med_occupation',
    'review_scores_rating',
    'reviews_per_month'
]

misc_cols = [
    'activity_months',
    'accommodates',
    'bathrooms', 
    'bedrooms',
    'cancellation_policy', 
    'cleaning_fee',
    'district',
    'extra_people',
    'first_review',
    'guests_included',
    'instant_bookable',
    'has_license',
    'host_response_time',
    'latitude',
    'longitude',
    'maximum_nights_avg_ntm',
    'minimum_nights_avg_ntm',
    'neighbourhood',
    'number_of_reviews',
    'number_of_reviews_ltm',
    'property_type',
    'room_type',
    'security_deposit',
    *key_cols
]

print_corr_map(df[misc_cols], height=900)

review_cols = [
    'activity_months',
    'instant_bookable',
    'review_scores_accuracy',
    'review_scores_cleanliness',
    'review_scores_checkin',
    'review_scores_communication',
    'review_scores_location',
    'review_scores_value',
    *key_cols
]

print_corr_map(df[review_cols])

### Licencia

La mayoría de las viviendas no presentan la licencia turística aunque no parece ser un asunto que penalice demasiado a la hora de ser alquilado en estos momentos.

Recientemente entró en vigor una [normativa municipal](https://elpais.com/ccaa/2019/03/27/madrid/1553675975_213545.html) para regular la actividad de alquileres vacacionales por lo que puede haber un período de cambios en este aspecto concreto que afecten a la actividad. 

Si la disponibilidad de datos lo permite será muy interesante evaluar cómo evoluciona este aspecto.

df_by_license = df.groupby(['has_license'])['has_license'].count().to_frame('has_license_count')
df_by_license.reset_index(inplace=True)

fig32 = go.Figure(
        go.Pie(
            labels=['no_license', 'has_license'], 
            values=df_by_license['has_license_count']
        )
)

fig32.update_traces(
    textfont_size=20,
    marker=dict(colors=['Orange', 'SteelBlue'])
)

fig32.update_layout(title='license')
fig32.show()

fig322 = go.Figure()
for val in [0, 1]:
    fig322.add_trace(
        go.Violin(
            x=df['has_license'][df['has_license'] == val],
            y=df['reviews_per_month'][df['has_license'] == val],
            name=val,                
            meanline_visible=True,
            line_color='Orange' if val == 0 else 'SteelBlue'
        )
    )

fig322.update_layout(title='reviews_per_month x has_license', showlegend=False)
fig322.show()

### Tipos de propiedad y relación con precio por huésped

Claro dominio de apartamentos. En el precio no hay diferencias notables.

df_by_property_type = df.groupby(['property_type'])['property_type'].count().to_frame('property_type_count')
df_by_property_type.reset_index(inplace=True)
property_types = np.sort(df_by_property_type['property_type'].unique())

fig33 = go.Figure(go.Pie(
    labels=df_by_property_type['property_type'], 
    values=df_by_property_type['property_type_count']
))

fig33.update_layout(title='property_type')
fig33.update_traces(textfont_size=15)
fig33.show()

fig332 = go.Figure()
for pt in property_types:
    fig332.add_trace(
        go.Violin(
            x=df['property_type'][df['property_type'] == pt],
            y=df['price_med_occupation_per_accommodate'][df['property_type'] == pt],
            points='all',
            name=pt,
            box_visible=True,                
            meanline_visible=True
        )
    )

fig332.update_layout(title='price_med_occupation_per_accommodate')
fig332.show()

### Tipos de habitaciones y relación con precio por huésped

El formato más habitual de alquiler vacacional es el piso completo, por encima de la habitación privada en piso compartido. Los precios van en sintonía con ello.

df_by_room_type = df.groupby(['room_type'])['room_type'].count().to_frame('room_type_count')
df_by_room_type.reset_index(inplace=True)
room_types = np.sort(df_by_room_type['room_type'].unique())

fig34 = go.Figure(
        go.Pie(
            labels=df_by_room_type['room_type'], 
            values=df_by_room_type['room_type_count']
        )
)

fig34.update_traces(textfont_size=20)
fig34.update_layout(title='room_type')
fig34.show()

fig342 = go.Figure()
for rt in room_types:
    fig342.add_trace(
        go.Violin(
            x=df['room_type'][df['room_type'] == rt],
            y=df['price_med_occupation_per_accommodate'][df['room_type'] == rt],
            points='all',
            name=rt,
            box_visible=True,                
            meanline_visible=True
        )
    )

fig342.update_layout(title='price_med_occupation_per_accommodate')
fig342.show()

### Relación entre número y puntuación de las reviews

Puede haber una cierta retroalimentación entre el número de reviews por mes y la nota en las mismas. Lo normal es que cuantas más y mejores reviews tenga una vivienda, más estancias genere.

fig27 = px.scatter(
    df, 
    x='review_scores_rating', 
    y='reviews_per_month', 
    color='room_type'
)

fig27.show()

### Relación entre política de cancelación y número de reviews mensuales

No parece que haya un claro impacto del tipo de cancelación sobre el número de estancias. En general los anfitriones utilizan políticas de cancelación que tienen un mínimo de flexibilidad en el tiempo para no ahuyentar a potenciales huéspedes.

df_by_cancellation_policy = df.groupby(['cancellation_policy'])['cancellation_policy'].count().to_frame('cancellation_policy_count')
df_by_cancellation_policy.reset_index(inplace=True)
cancellation_policy_types = np.sort(df_by_cancellation_policy['cancellation_policy'].unique())

fig36 = go.Figure(go.Pie(
    labels=df_by_cancellation_policy['cancellation_policy'], 
    values=df_by_cancellation_policy['cancellation_policy_count']
))

fig36.update_layout(title='cancellation_policy')
fig36.update_traces(textfont_size=15)
fig36.show()

fig362 = go.Figure()
for cpt in cancellation_policy_types:
    fig362.add_trace(
        go.Violin(
            x=df['cancellation_policy'][df['cancellation_policy'] == cpt],
            y=df['reviews_per_month'][df['cancellation_policy'] == cpt],
            name=cpt,
            box_visible=True,                
            meanline_visible=True
        )
    )

fig362.update_layout(title='reviews_per_month', showlegend=False)
fig362.show()

### Relación entre número mínimo de noches y número de reviews mensuales

El número mínimo de noches a la hora de reservar un alojamiento podría ser un factor limitante. Simplemente es un indicio porque no se conoce la duración de las estancias pero los alojamientos que menos limitan este aspecto tienen más estancias al mes.

fig29 = px.scatter(
    df, 
    x='minimum_nights_avg_ntm', 
    y='reviews_per_month'
)

fig29.show()

### Relación entre tiempo de respuesta y número de reviews mensuales

A mejor tasa de respuesta, más estancias. Es importante por tanto la agilidad de los anfitriones a la hora de tratar con los potenciales huéspedes.

df['host_response_time'].fillna('-unk-', inplace=True)
df_by_response_time = df.groupby(['host_response_time'])['host_response_time'].count().to_frame('host_response_time_count')
df_by_response_time.reset_index(inplace=True)
response_time_types = np.sort(df_by_response_time['host_response_time'].unique())

fig38 = go.Figure()
for rtt in response_time_types:
    fig38.add_trace(
        go.Violin(
            x=df['host_response_time'][df['host_response_time'] == rtt],
            y=df['reviews_per_month'][df['host_response_time'] == rtt],
            name=rtt,
            box_visible=True,                
            meanline_visible=True
        )
    )

fig38.update_layout(title='reviews_per_month', showlegend=False)
fig38.show()

### Tasa de limpieza

Una mayor tasa de limpieza se relaciona con viviendas que se alquilan de forma completa. Cuanto más alta es la tasa de limpieza, menos reservas al mes se obtienen posiblemente por tratarse de viviendas exclusivas.

fig211 = px.scatter(
    df, 
    x='cleaning_fee', 
    y='reviews_per_month', 
    color='room_type'
)

fig211.show()

fig2112 = px.scatter(
    df, 
    x='cleaning_fee', 
    y='income_med_occupation', 
    color='room_type'
)

fig2112.show()

### Número de alojamientos por barrio

El centro tiene una concentración mucho más alta de alojamientos.

df_by_nb = df.groupby(['neighbourhood'])['neighbourhood'].size().to_frame('count')
df_by_nb.reset_index(inplace=True)

fig310 = go.Figure(go.Choroplethmapbox(
    geojson=city_nb,
    locations=df_by_nb['neighbourhood'], 
    z=df_by_nb['count'],                   
    colorscale='Blues',                                
    marker_opacity=0.5, 
    marker_line_width=0.5
))

fig310.update_layout(
    mapbox_style='carto-positron',
    mapbox_zoom=11, 
    mapbox_center={'lat':df['latitude'].mean(), 'lon':df['longitude'].mean()},
    margin={"r":0,"t":0,"l":0,"b":0}
)

fig310.show()

### Relación entre precio por húesped y localización

En general son los alojamientos de los barrios de los distritos Moncloa, Centro, Salamanca y Chamartín los que presentan precios más altos. Puede haber excepciones por diferentes motivos.

df_by_nb_p = df.groupby(['neighbourhood', 'district'])['price_med_occupation_per_accommodate'].mean().to_frame('nb_avg')
df_by_nb_p.reset_index(inplace=True)
print(df_by_nb_p.sort_values(by='nb_avg', ascending=False).head(15))

fig311 = go.Figure(
    go.Scattermapbox(
        lon=df['longitude'],
        lat=df['latitude'],
        mode='markers',
        marker_color=df['price_med_occupation_per_accommodate'],
        text=df['price_med_occupation_per_accommodate'],
        marker=dict(
            opacity=0.5, 
            colorscale='Blues',
            cmin=df['price'].min(), 
            cmax=df['price'].max()
        )
    )
)

fig311.update_layout(
    mapbox_style='carto-positron',
    mapbox_zoom=11, 
    mapbox_center={'lat':df['latitude'].mean(), 'lon':df['longitude'].mean()},
    margin={"r":0,"t":0,"l":0,"b":0}
)

fig311.show()

### Relación entre precio por huésped y notas de review

Las notas de las reviews no tienen una relación clara con el precio pagado por el huésped (ya sea alto o bajo).

fig214 = px.scatter(
    df, 
    x='review_scores_rating',
    y='price_med_occupation_per_accommodate'
)

fig214.show()

### Relación entre precio por huésped y número de reviews mensuales

Los alojamientos con precios por noche en el rango entre 100 y 170 euros son los que consiguen más estancias.

fig215 = px.scatter(
    df,
    x='price_med_occupation_per_accommodate',
    y='reviews_per_month',
    marginal_x='box', 
    trendline='ols'
)

fig215.show()

### Review de localización
La valoración de localización es un aspecto muy abierto e interesante. Se puede considerar una localización en función de si está bien conectada con el resto de la ciudad, o de si está próxima a puntos de interés, o de si está bien ubicado con respecto a cualquier intención concreta ya sea de tipo turístico, laboral, familiar, etc. que cualquier huésped puede tener.

En un primer mapa se observa que la calificación de la review de localización no tiene una relación destacable con la cercanía a puntos de transporte como Metro o Cercanías (puntos de color negro) sino que más bien la tiene con la cercanía al centro, sin ser esta una tendencia definitiva (la calificación de review se pinta en escala de colores cálidos, decreciendo la nota del amarillo al rojo).

En el segundo mapa se muestra que efectivamente los barrios del centro tienen una gran calificación en este aspecto pero tampoco hay diferencias muy grandes con barrios un poco más lejanos al centro.

fig314 = go.Figure()

df_location = df[~df['review_scores_location'].isnull()]

fig314.add_trace(go.Scattermapbox(
    lon=df_location['longitude'],
    lat=df_location['latitude'],
    mode='markers',
    marker_color=df_location['review_scores_location'],
    text=df_location['review_scores_location'],
    marker=dict(
        size=5,
        opacity=0.5,
        colorscale='YlOrRd',
        reversescale=True,
        cmin=0, 
        cmax=10
    )
))

fig314.add_trace(
    go.Scattermapbox(
        lon=dfpt['lon'],
        lat=dfpt['lat'],
        mode='markers',
        text=dfpt['type'] + ' ' + dfpt['name'],
        marker=dict(
            size=5,
            opacity=0.7,
            color='black'
        )
    )
)

fig314.update_layout(
    showlegend=False,
    mapbox_style='carto-positron',
    mapbox_zoom=11, 
    mapbox_center={'lat':df['latitude'].mean(), 'lon':df['longitude'].mean()},
    margin={"r":0,"t":0,"l":0,"b":0}
)

fig314.show()

df_by_nb = df.groupby(['neighbourhood'])['review_scores_location'].mean().to_frame('review_scores_location_avg')
df_by_nb.reset_index(inplace=True)

fig3142 = go.Figure(go.Choroplethmapbox(
    geojson=city_nb,
    locations=df_by_nb['neighbourhood'], 
    z=df_by_nb['review_scores_location_avg'],                   
    colorscale='Blues',                                
    marker_opacity=0.5, 
    marker_line_width=0.5
))

fig3142.update_layout(
    mapbox_style='carto-positron',
    mapbox_zoom=11, 
    mapbox_center={'lat':df['latitude'].mean(), 'lon':df['longitude'].mean()},
    margin={"r":0,"t":0,"l":0,"b":0}
)

fig3142.show()

### Distribución del precio por huésped

Entre 60 y 150 euros por noche es el rango más habitual de precio que paga un huésped por noche por una estancia de duración media.

fig = px.histogram(df, x='price_med_occupation_per_accommodate', nbins=40)
fig.show()

fig = go.Figure()

fig.add_trace(go.Histogram(
    x=df['price_med_occupation_per_accommodate'],
    name='price_med_occupation_per_accommodate',
    marker_color='#EB89B5'))

fig.add_trace(go.Histogram(
    x=df['price'],
    name='price',
    marker_color='#330C73'))

"""
fig.add_trace(go.Histogram(
    x=df['income_med_occupation'],
    name='income_med_occupation',
    marker_color='green'))
"""

fig.update_layout(barmode='stack')
fig.update_traces(opacity=0.75)
fig.show()

### Relación entre habitaciones y baños

Dominan con claridad los alojamientos con pocas estancias, en concreto con un baño y un dormitorio.

plt.figure(figsize=(16, 8))
plt.hist(df['bedrooms'], bins=15)
plt.gca().set(title='BEDROOMS', ylabel='COUNT');

df['bathrooms'].fillna(0, inplace=True)
plt.figure(figsize=(16, 8))
plt.hist(df['bathrooms'], bins=15)
plt.gca().set(title='BATHROOMS', ylabel='COUNT');

df_bb = df.groupby(['bedrooms', 'bathrooms'])[['bedrooms', 'bathrooms']].size().to_frame('cnt')
df_bb.reset_index(inplace=True)
fig316 = px.scatter(df_bb, x='bedrooms', y='bathrooms', size='cnt')
fig316.show()

### Relación entre ingresos por estancia y capacidad

Tendencia creciente de los ingresos según la capacidad del alojamiento. Es interesante cómo también influye el tamaño de los alojamientos, por ejemplo en los distritos Centro, Chamartín, Salamanca, Retiro o Moncloa las capacidades de las viviendas son mucho mayores que por ejemplo en Latina.

df_by_price_per_bed = df.groupby(['district', 'accommodates'])['income_med_occupation'].mean().to_frame('income_med_occupation_avg')
df_by_price_per_bed.reset_index(inplace=True)

px.line(
    df_by_price_per_bed, 
    x='accommodates', 
    y='income_med_occupation_avg',
    color='district'
).show()

### Distribución de ingresos por estancia según el distrito

Estas distribuciones refuerzan el análisis del apartado anterior.

districts = np.sort(df['district'].unique())[::-1]      
    
fig318 = go.Figure()
for d in districts:
    fig318.add_trace(go.Violin(
        x=df[df['district'] == d]['income_med_occupation'].values,
        name='  ' + d + '  '
    ))

fig318.update_traces(
    orientation='h', 
    side='positive', 
    width=3, 
    points=False
)

fig318.update_layout(
    height=900,
    xaxis_showgrid=False, 
    xaxis_zeroline=False, 
    showlegend=False
)

fig318.show()

fig220 = px.scatter(
    df,
    x='accommodates',
    y='income_med_occupation', 
    trendline='ols',
    color='district'
)

fig220.show()

### Relación entre barrio y precio por huésped

En general son los barrios más céntricos los que mayor precio por huésped tienen aunque hay diversas excepciones.

Los barrios del sur en general presentan precios más bajos que el resto.

df_by_nb = df.groupby(['neighbourhood'])['price_med_occupation_per_accommodate'].mean().to_frame('price_med_occupation_per_accommodate_avg')
df_by_nb.reset_index(inplace=True)

fig320 = go.Figure(go.Choroplethmapbox(
    geojson=city_nb,
    locations=df_by_nb['neighbourhood'], 
    z=df_by_nb['price_med_occupation_per_accommodate_avg'],                   
    colorscale='Blues',                                
    marker_opacity=0.5, 
    marker_line_width=0.5
))

fig320.update_layout(
    mapbox_style='carto-positron',
    mapbox_zoom=11, 
    mapbox_center={'lat':df['latitude'].mean(), 'lon':df['longitude'].mean()},
    margin={"r":0,"t":0,"l":0,"b":0}
)

fig320.show()

### Relación entre distrito y precio medio de alquiler por m2

Los distritos del centro tienen los precios medios de alquiler más elevados.

df_by_district = df.groupby(['district'])['income_med_occupation'].mean().to_frame('income_med_occupation_avg')
df_by_district.reset_index(inplace=True)
df_by_district = df_by_district.merge(df_price_rent_m2) # join

fig321 = go.Figure(go.Choroplethmapbox(
    geojson=city_districts,
    locations=df_by_district['district'], 
    z=df_by_district['price_rent_eur_m2'],                   
    colorscale='Blues',                                
    marker_opacity=0.5, 
    marker_line_width=0.5
))

fig321.update_layout(
    mapbox_style='carto-positron',
    mapbox_zoom=11, 
    mapbox_center={'lat':df['latitude'].mean(), 'lon':df['longitude'].mean()},
    margin={"r":0,"t":0,"l":0,"b":0}
)

fig321.show()

### Relación entre precio por huésped y precio medio de alquiler por m2

Los precios por huésped están bastante alineados con el precio de alquiler que presenta el mercado inmobiliario.

df_by_district = df.groupby(['district'])['price_med_occupation_per_accommodate'].mean().to_frame('price_med_occupation_per_accommodate_avg')
df_by_district.reset_index(inplace=True)
df_by_district = df_by_district.merge(df_price_rent_m2) # join

fig322 = px.scatter(
    df_by_district, 
    x='price_rent_eur_m2',
    y='price_med_occupation_per_accommodate_avg'
)

fig322.show()

### Relación entre precio por huésped e ingresos por estancia media

A lo largo del estudio se han presentado diferentes relaciones entre características donde participan el precio medio por noche y huésped y los ingresos por estancia media. Ambas variables están basadas en la característica precio y por ello tienen una clara relación con ella así que es probable que en la fase de modelado se descarte el uso de alguna de ellas.

fig = go.Figure()

fig.add_trace(go.Histogram(
    x=df['price_med_occupation_per_accommodate'],
    name='price_med_occupation_per_accommodate',
    marker_color='#EB89B5'))

fig.add_trace(go.Histogram(
    x=df['income_med_occupation'],
    name='income_med_occupation',
    marker_color='blue'))

fig.update_layout(barmode='stack')
fig.update_traces(opacity=0.75)
fig.show()

fig323 = go.Figure()
for x in ['price_med_occupation_per_accommodate', 'income_med_occupation']:
    fig323.add_trace(go.Scatter(
        x=df['price'], 
        y=df[x], 
        mode='markers',
        name=x,
        marker=dict(color=('blue' if x != 'price_med_occupation_per_accommodate' else '#EB89B5'))
    ))
    fig323.update_traces(marker=dict(opacity=0.5))

fig323.show()

### Equipamiento de la vivienda

En las correlaciones simplemente se observa más afinidad entre objetos que normalmente se ubican en una cocina como el frigorífico, el microondas, la cafetera o la vajilla.

No existe un impacto claro entre alguna comodidad concreta y las reviews en número o nota de las estancias.

Curiosamente el aire acondicionado es una de las comodidades que menos destaca, bien es cierto que es algo que tiene importancia sobre todo de forma estacional y según para qué tipo de huéspedes.

reviews_cols = [
    'reviews_per_month', 
    'review_scores_rating'
]

amenities_cols = [
    'has_air_conditioning',
    'has_bed_linens',
    'has_coffee_maker',
    'has_cooking_basics',
    'has_dishes_and_silverware',
    'has_elevator',
    'has_essentials', 
    'has_family/kid_friendly',
    'has_first_aid_kit',
    'has_hair_dryer',
    'has_hangers',
    'has_heating',
    'has_hot_water',
    'has_iron',
    'has_kitchen',
    'has_laptop_friendly_workspace',
    'has_microwave',
    'has_no_stairs_or_steps_to_enter',
    'has_oven',
    'has_refrigerator',
    'has_shampoo',
    'has_stove',
    'has_tv',
    'has_washer',
    'has_wifi'
]

print_corr_map(df[[*amenities_cols, *reviews_cols]], height=900)

"""
for am in amenities_cols:
    fig324 = go.Figure()
    fig324 = make_subplots(
        rows=1, 
        cols=2, 
        subplot_titles=('reviews_per_month', 'review_scores_rating')
    )
    
    for var in reviews_cols:
        for val in [0, 1]:
            fig324.add_trace(go.Violin(
                    x=df[am][df[am] == val],
                    y=df[var][df[am] == val],
                    name=val,                
                    meanline_visible=True,
                    line_color='Orange' if val == 0 else 'SteelBlue'
                ),
                row=1, 
                col=1 if var == 'reviews_per_month' else 2
            )

        fig324.update_layout(title=am, showlegend=False)
        
    fig324.show()
"""

## Exportación dataset

df.to_csv(filename_out, index=False)
